fix park details layout. update/add comment in UI in db for rating

+++ app/src/main/java/com/psu/sweng500/team4/parkpal/Models/Weather/Wind.java
@@ -13,7 +13,7 @@ public class Wind {
     private Double speed;
     @SerializedName("deg")
     @Expose
-    private Integer deg;
+    private Float deg;
 
     public Double getSpeed() {
         return speed;
@@ -23,11 +23,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public Integer getDeg() {
+    public Float getDeg() {
         return deg;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Float deg) {
         this.deg = deg;
     }
